#412 rename range module to avoid clash with Python built-in
